requirements.txt
# Core
duckdb~=1.3.1
numpy~=2.3.1
pandas~=2.3.0
sqlglot~=26.33
tabulate~=0.9.0
requests~=2.32.4
tqdm~=4.67.1
beartype~=0.21.0

# Graphs
rdflib~=7.1.4
SPARQLWrapper~=2.0.0
oxrdflib~=0.4.0

# Models
nltk~=3.9.1
rank-bm25~=0.2.2
azure-ai-ml~=1.27.1
azure-identity~=1.17.1
azure-ai-inference~=1.0.0b9
openai~=1.107.3
google-genai~=1.73.1
anthropic~=0.85.0
gitpython==3.1.46

# Misc.
pytest~=8.4.1
python-json-logger~=3.3.0
matplotlib~=3.10.7
plotly~=6.3.1
